Add Docs for Dialog `template` option (#11912)
