+++ docs/content/guides/dialog/dialog/angular/example5.html
