+++ docs/content/guides/dialog/dialog/angular/example5.ts
@@ -5,9 +5,19 @@ import { GridSettings, HotTableComponent } from '@handsontable/angular-wrapper';
 @Component({
   selector: 'app-example5',
   template: `
+    <div class="example-controls-container" style="padding-bottom: 16px">
+      <div class="controlsQuickFilter">
+        <label for="background-select" class="selectColumn">Select a background:
+          <select id="background-select" (change)="onBackgroundChange($event)">
+            <option value="solid">Solid</option>
+            <option value="semi-transparent">Semi-transparent</option>
+          </select>
+        </label>
+      </div>
+    </div>
     <hot-table
       #hotTable
-      [settings]="hotSettings!" 
+      [settings]="hotSettings!"
       [data]="hotData"
     >
     </hot-table>
@@ -100,9 +110,8 @@ export class AppComponent implements AfterViewInit {
     autoWrapCol: true,
     autoRowSize: true,
     dialog: {
-      content: 'This dialog uses a semi-transparent and content background.',
-      contentBackground: true,
-      background: 'semi-transparent',
+      content: 'This dialog uses a solid background (default).',
+      background: 'solid',
       closable: true,
     },
   };
@@ -114,6 +123,21 @@ export class AppComponent implements AfterViewInit {
       // Show dialog after initialization
       const dialogPlugin = hotInstance.getPlugin('dialog');
       dialogPlugin.show();
+    }
+  }
+
+  onBackgroundChange(event: Event) {
+    const hotInstance = this.hotTable.hotInstance;
+
+    if (hotInstance) {
+      const background = (event.target as HTMLSelectElement).value;
+      const content = background === 'solid' ? 'This dialog uses a solid background (default).' : 'This dialog uses a semi-transparent background.';
+      const dialogPlugin = hotInstance.getPlugin('dialog');
+
+      dialogPlugin.update({
+        content,
+        background,
+      });
     }
   }
 }
