+++ docs/content/guides/dialog/dialog/angular/example6.html
