+++ docs/content/guides/dialog/dialog/angular/example7.html
