+++ docs/content/guides/dialog/dialog/angular/example8.html
@@ -1,0 +1,3 @@
+<div>
+  <app-example8></app-example8>
+</div>
