+++ docs/content/guides/dialog/dialog/javascript/example4.html
@@ -1,11 +1,6 @@
-<div class="example-controls-container" style="padding-bottom: 16px">
-  <div class="controlsQuickFilter">
-    <label for="background-select" class="selectColumn">Select a background:
-      <select id="background-select">
-        <option value="solid">Solid</option>
-        <option value="semi-transparent">Semi-transparent</option>
-      </select>
-    </label>
-  </div>
+<div style="margin-bottom: 16px; display: flex; gap: 10px">
+  <button id="showAlert">Show Alert</button>
+  <button id="showConfirm">Show Confirm</button>
+  <button id="showCustomConfirm">Show custom Confirm (with solid background)</button>
 </div>
 <div id="example4"></div>
