+++ docs/content/guides/dialog/dialog/javascript/example4.js
@@ -59,7 +59,7 @@ const hot = new Handsontable(container, {
       title: 'Date',
       type: 'date',
       data: 'sellDate',
-      width: 130,
+      width: 131,
       dateFormat: 'MMM D, YYYY',
       correctFormat: true,
       className: 'htRight',
@@ -86,28 +86,80 @@ const hot = new Handsontable(container, {
   width: '100%',
   height: 300,
   stretchH: 'all',
-  dialog: {
-    content: 'This dialog uses a solid background (default).',
-    background: 'solid',
-    closable: true,
-  },
+  dialog: true,
   licenseKey: 'non-commercial-and-evaluation',
 });
 
-// Show dialog after initialization
 const dialogPlugin = hot.getPlugin('dialog');
 
-dialogPlugin.show();
-// Add event listeners for select
-document.getElementById('background-select').addEventListener('change', (event) => {
-  const background = event.target.value;
-  const content =
-    background === 'solid'
-      ? 'This dialog uses a solid background (default).'
-      : 'This dialog uses a semi-transparent background.';
-
-  dialogPlugin.update({
-    content,
-    background,
+// Add event listeners for buttons
+document.getElementById('showAlert').addEventListener('click', () => {
+  dialogPlugin.showAlert(
+    {
+      title: 'Alert',
+      description: 'This is an example of the alert dialog.',
+    },
+    () => {
+      dialogPlugin.hide();
+    }
+  );
+});
+document.getElementById('showConfirm').addEventListener('click', () => {
+  dialogPlugin.showConfirm(
+    'Do you want to undo the last action?',
+    () => {
+      hot.getPlugin('undoRedo').undo();
+      dialogPlugin.hide();
+    },
+    () => {
+      dialogPlugin.hide();
+    }
+  );
+});
+document.getElementById('showCustomConfirm').addEventListener('click', () => {
+  dialogPlugin.show({
+    template: {
+      type: 'confirm',
+      title: 'Increase the price of the first row by:',
+      buttons: [
+        {
+          text: 'Cancel',
+          type: 'secondary',
+          callback: () => {
+            dialogPlugin.hide();
+          },
+        },
+        {
+          text: '$100',
+          type: 'secondary',
+          callback: () => {
+            dialogPlugin.hide();
+            hot.setDataAtCell(0, 1, hot.getDataAtCell(0, 1) + 100);
+            hot.selectCell(0, 1);
+          },
+        },
+        {
+          text: '$200',
+          type: 'secondary',
+          callback: () => {
+            dialogPlugin.hide();
+            hot.setDataAtCell(0, 1, hot.getDataAtCell(0, 1) + 200);
+            hot.selectCell(0, 1);
+          },
+        },
+        {
+          text: '$500',
+          type: 'secondary',
+          callback: () => {
+            dialogPlugin.hide();
+            hot.setDataAtCell(0, 1, hot.getDataAtCell(0, 1) + 500);
+            hot.selectCell(0, 1);
+          },
+        },
+      ],
+    },
+    background: 'solid',
+    contentBackground: false,
+    closable: true,
   });
 });
