+++ docs/content/guides/dialog/dialog/javascript/example5.html
@@ -1,0 +1,11 @@
+<div class="example-controls-container" style="padding-bottom: 16px">
+  <div class="controlsQuickFilter">
+    <label for="background-select" class="selectColumn">Select a background:
+      <select id="background-select">
+        <option value="solid">Solid</option>
+        <option value="semi-transparent">Semi-transparent</option>
+      </select>
+    </label>
+  </div>
+</div>
+<div id="example5"></div>
