+++ docs/content/guides/dialog/dialog/javascript/example5.ts
@@ -60,7 +60,7 @@ const hot = new Handsontable(container, {
       title: 'Date',
       type: 'date',
       data: 'sellDate',
-      width: 130,
+      width: 131,
       dateFormat: 'MMM D, YYYY',
       correctFormat: true,
       className: 'htRight',
@@ -88,9 +88,8 @@ const hot = new Handsontable(container, {
   height: 300,
   stretchH: 'all',
   dialog: {
-    content: 'This dialog uses a semi-transparent background.',
-    contentBackground: true,
-    background: 'semi-transparent',
+    content: 'This dialog uses a solid background (default).',
+    background: 'solid',
     closable: true,
   },
   licenseKey: 'non-commercial-and-evaluation',
@@ -100,3 +99,17 @@ const hot = new Handsontable(container, {
 const dialogPlugin = hot.getPlugin('dialog');
 
 dialogPlugin.show();
+
+// Add event listeners for select
+(document.getElementById('background-select') as HTMLSelectElement).addEventListener('change', (event) => {
+  const background = (event.target as HTMLSelectElement).value;
+  const content =
+    background === 'solid'
+      ? 'This dialog uses a solid background (default).'
+      : 'This dialog uses a semi-transparent background.';
+
+  dialogPlugin.update({
+    content,
+    background,
+  });
+});
