+++ docs/content/guides/dialog/dialog/javascript/example6.ts
@@ -60,7 +60,7 @@ const hot = new Handsontable(container, {
       title: 'Date',
       type: 'date',
       data: 'sellDate',
-      width: 130,
+      width: 131,
       dateFormat: 'MMM D, YYYY',
       correctFormat: true,
       className: 'htRight',
@@ -88,13 +88,9 @@ const hot = new Handsontable(container, {
   height: 300,
   stretchH: 'all',
   dialog: {
-    content: '<h2 id="example6-title">Title</h2><p id="example6-description">Description</p>',
-    a11y: {
-      role: 'alertdialog',
-      ariaLabel: 'Title',
-      ariaLabelledby: 'example6-title',
-      ariaDescribedby: 'example6-description',
-    },
+    content: 'This dialog uses a semi-transparent background.',
+    contentBackground: true,
+    background: 'semi-transparent',
     closable: true,
   },
   licenseKey: 'non-commercial-and-evaluation',
