+++ docs/content/guides/dialog/dialog/javascript/example8.html
@@ -2,4 +2,4 @@
   <button id="showDialog">Show Dialog</button>
   <button id="hideDialog">Hide Dialog</button>
 </div>
-<div id="example7"></div>
+<div id="example8"></div>
