+++ docs/content/guides/dialog/dialog/react/example1.tsx
@@ -77,7 +77,7 @@ const ExampleComponent = () => {
         title="Date"
         type="date"
         data="sellDate"
-        width={130}
+        width={131}
         dateFormat="MMM D, YYYY"
         correctFormat={true}
         className="htRight"
