+++ docs/content/guides/dialog/dialog/react/example3.jsx
@@ -81,7 +81,7 @@ const ExampleComponent = () => {
         title="Date"
         type="date"
         data="sellDate"
-        width={130}
+        width={131}
         dateFormat="MMM D, YYYY"
         correctFormat={true}
         className="htRight"
